requirements.txt
singleton-decorator==1.0.0 
mysql-connector-python==8.0.20